refactor(router-store): refactor router_store_module into separate files (#1299)
Closes #1298
