+++ modules/router-store/src/index.ts
@@ -2,28 +2,28 @@ export {
   ROUTER_ERROR,
   ROUTER_CANCEL,
   ROUTER_NAVIGATION,
-  ROUTER_REQUEST,
   ROUTER_NAVIGATED,
-  RouterNavigationAction,
-  RouterCancelAction,
-  RouterErrorAction,
-  RouterRequestAction,
-  RouterNavigatedAction,
+  ROUTER_REQUEST,
   RouterAction,
-  routerReducer,
-  RouterErrorPayload,
-  RouterReducerState,
+  RouterCancelAction,
   RouterCancelPayload,
-  RouterNavigationPayload,
-  RouterRequestPayload,
+  RouterErrorAction,
+  RouterErrorPayload,
+  RouterNavigatedAction,
   RouterNavigatedPayload,
+  RouterNavigationAction,
+  RouterNavigationPayload,
+  RouterRequestAction,
+  RouterRequestPayload,
+} from './actions';
+export { routerReducer, RouterReducerState } from './reducer';
+export {
   StoreRouterConnectingModule,
   StoreRouterConfig,
   NavigationActionTiming,
   ROUTER_CONFIG,
   DEFAULT_ROUTER_FEATURENAME,
 } from './router_store_module';
-
 export {
   RouterStateSerializer,
   DefaultRouterStateSerializer,
