+++ modules/router-store/src/router_store_module.ts
@@ -14,166 +14,21 @@ import {
   NavigationStart,
 } from '@angular/router';
 import { select, Store } from '@ngrx/store';
+import { withLatestFrom } from 'rxjs/operators';
 
+import {
+  ROUTER_CANCEL,
+  ROUTER_ERROR,
+  ROUTER_NAVIGATED,
+  ROUTER_NAVIGATION,
+  ROUTER_REQUEST,
+} from './actions';
+import { RouterReducerState } from './reducer';
 import {
   DefaultRouterStateSerializer,
   RouterStateSerializer,
   SerializedRouterStateSnapshot,
-  BaseRouterStoreState,
 } from './serializer';
-import { withLatestFrom } from 'rxjs/operators';
-
-/**
- * An action dispatched when a router navigation request is fired.
- */
-export const ROUTER_REQUEST = 'ROUTER_REQUEST';
-
-/**
- * Payload of ROUTER_REQUEST
- */
-export type RouterRequestPayload = {
-  event: NavigationStart;
-};
-
-/**
- * An action dispatched when a router navigation request is fired.
- */
-export type RouterRequestAction = {
-  type: typeof ROUTER_REQUEST;
-  payload: RouterRequestPayload;
-};
-
-/**
- * An action dispatched when the router navigates.
- */
-export const ROUTER_NAVIGATION = 'ROUTER_NAVIGATION';
-
-/**
- * Payload of ROUTER_NAVIGATION.
- */
-export type RouterNavigationPayload<T extends BaseRouterStoreState> = {
-  routerState: T;
-  event: RoutesRecognized;
-};
-
-/**
- * An action dispatched when the router navigates.
- */
-export type RouterNavigationAction<
-  T extends BaseRouterStoreState = SerializedRouterStateSnapshot
-> = {
-  type: typeof ROUTER_NAVIGATION;
-  payload: RouterNavigationPayload<T>;
-};
-
-/**
- * An action dispatched when the router cancels navigation.
- */
-export const ROUTER_CANCEL = 'ROUTER_CANCEL';
-
-/**
- * Payload of ROUTER_CANCEL.
- */
-export type RouterCancelPayload<T, V extends BaseRouterStoreState> = {
-  routerState: V;
-  storeState: T;
-  event: NavigationCancel;
-};
-
-/**
- * An action dispatched when the router cancel navigation.
- */
-export type RouterCancelAction<
-  T,
-  V extends BaseRouterStoreState = SerializedRouterStateSnapshot
-> = {
-  type: typeof ROUTER_CANCEL;
-  payload: RouterCancelPayload<T, V>;
-};
-
-/**
- * An action dispatched when the router errors.
- */
-export const ROUTER_ERROR = 'ROUTE_ERROR';
-
-/**
- * Payload of ROUTER_ERROR.
- */
-export type RouterErrorPayload<T, V extends BaseRouterStoreState> = {
-  routerState: V;
-  storeState: T;
-  event: NavigationError;
-};
-
-/**
- * An action dispatched when the router errors.
- */
-export type RouterErrorAction<
-  T,
-  V extends BaseRouterStoreState = SerializedRouterStateSnapshot
-> = {
-  type: typeof ROUTER_ERROR;
-  payload: RouterErrorPayload<T, V>;
-};
-
-/**
- * An action dispatched after navigation has ended and new route is active.
- */
-export const ROUTER_NAVIGATED = 'ROUTER_NAVIGATED';
-
-/**
- * Payload of ROUTER_NAVIGATED.
- */
-export type RouterNavigatedPayload = {
-  event: NavigationEnd;
-};
-
-/**
- * An action dispatched after navigation has ended and new route is active.
- */
-export type RouterNavigatedAction = {
-  type: typeof ROUTER_NAVIGATED;
-  payload: RouterNavigatedPayload;
-};
-
-/**
- * An union type of router actions.
- */
-export type RouterAction<
-  T,
-  V extends BaseRouterStoreState = SerializedRouterStateSnapshot
-> =
-  | RouterRequestAction
-  | RouterNavigationAction<V>
-  | RouterCancelAction<T, V>
-  | RouterErrorAction<T, V>
-  | RouterNavigatedAction;
-
-export type RouterReducerState<
-  T extends BaseRouterStoreState = SerializedRouterStateSnapshot
-> = {
-  state: T;
-  navigationId: number;
-};
-
-export function routerReducer<
-  T extends BaseRouterStoreState = SerializedRouterStateSnapshot
->(
-  state: RouterReducerState<T> | undefined,
-  action: RouterAction<any, T>
-): RouterReducerState<T> {
-  switch (action.type) {
-    case ROUTER_NAVIGATION:
-    case ROUTER_ERROR:
-    case ROUTER_CANCEL:
-      return {
-        state: action.payload.routerState,
-        navigationId: action.payload.event.id,
-      };
-    default:
-      return state as RouterReducerState<T>;
-  }
-}
 
 export interface StoreRouterConfig {
   stateKey?: string;
@@ -314,7 +169,10 @@ export class StoreRouterConnectingModule {
 
   private setUpStoreStateListener(): void {
     this.store
-      .pipe(select(this.stateKey), withLatestFrom(this.store))
+      .pipe(
+        select(this.stateKey),
+        withLatestFrom(this.store)
+      )
       .subscribe(([routerStoreState, storeState]) => {
         this.navigateIfNeeded(routerStoreState, storeState);
       });
